refactor: #96 moved feedback service to shared folder
